+++ packages/perseus-editor/src/widgets/interactive-graph-editor/components/interactive-graph-settings.test.tsx
@@ -3,8 +3,8 @@ import {render, screen, waitFor, fireEvent} from "@testing-library/react";
 import {userEvent as userEventLib} from "@testing-library/user-event";
 import * as React from "react";
 
-import {testDependencies} from "../../../../../testing/test-dependencies";
-import InteractiveGraphSettings from "../../widgets/interactive-graph-editor/components/interactive-graph-settings";
+import {testDependencies} from "../../../../../../testing/test-dependencies";
+import InteractiveGraphSettings from "./interactive-graph-settings";
 
 import type {UserEvent} from "@testing-library/user-event";
